fss: carry FSS altitude at full width so only the clamp narrows it (todo/55)
The FSS wire altitude (fss_message_asset_command::getAltitude, uint32_t feet)
was narrowed to uint16_t in FSSCommandTarget before the [floor, cap] clamp ever
saw it. A malformed or oversized value could wrap during that cast to a small
altitude the safety clamp would then accept as legitimate -- e.g. 65536 ft
wrapping to 0, clamped up to the floor, so a "far too high" command became "as
low as allowed".

Widen the path (FSSCommandTarget::altitude, IMAV/MAV::setAltitude,
mav_connection::commandAltitude, clamp_command_altitude) to uint32_t so
clamp_command_altitude is the single narrowing authority: an oversized value
converts to a large metre count and pins to the cap, never wraps.

Tests: an oversized altitude reaches setAltitude intact through the state
machine, and clamp_command_altitude pins 65536/65569/UINT32_MAX to the cap.

+++ src/fss/fmu-fss-types.hpp
@@ -25,7 +25,12 @@ enum FSSCommand
 struct FSSCommandTarget
 {
     Point position{};
-    uint16_t altitude{ 0 };
+    /* Feet, as delivered on the FSS wire (fss_message_asset_command::getAltitude
+     * is uint32_t). Kept full-width here so the value is narrowed only by the
+     * regulatory [floor, cap] clamp in clamp_command_altitude(), never by a lossy
+     * cast that could wrap a large altitude down to a low one before the clamp
+     * ever sees it (todo/55). */
+    uint32_t altitude{ 0 };
 };
 
 enum FSSCommsStatus
